Render title text (#2)
* feat: add text component

* feat: use pattern matcher in makefile

With a pattern matcher, new cpp files that need to be compiled do not
need their own rules added to the file. Now I don't need to manually
maintain the makefile as much. Compilation of main.cpp is still written
manually, but that one currently is pretty custom and I don't know of a
benefit yet of making that one more automated at the moment.

* feat: add text component system; render text components each frame

* feat: render title text & start text

* chore: remove unused DemoTimerComponent