feat: add input system
refactor: add main menu class

The main menu code that creates entities to render itself can now
conveniently be found in a main menu class.

feat: dispatch normal events from sfml window events

Now, when sfml input events come in through the window object, event
objects are fired via the event Dispatcher to anyone who's listening.
This means that subscribers to input events do not need a reference to
the sfml window object.

A cost of this approach is that the input events are received by all
subscribers, instead of being consumed by the highest priority
subscriber. So potentially two subscribers could perform actions at the
same time if they are listening to the same input event, but that's an
okay cost for this game. I don't think there will be a time when two
different systems are subscribed to the same input event.